requirements.txt
autopep8==1.6.0
beautifulsoup4==4.10.0
lxml==4.6.4
pandas==1.3.4
python-whois==0.7.3
requests==2.22.0
